Default template of a schema description

Default template of a schema description

schema_default

Format

The object of class schema describes at which position in a table which information can be found. It contains the four slots clusters, header, meta and variables.

There is hardly any limit to how data can be arranged in a spreadsheet, apart from the apparent organisation into a lattice of cells. However, it is often the case that data are gathered into topologically coherent chunks. Those chunks are what is considered 'cluster' in arealDB. Clusters are described by the properties:

row [integerish(1)]:
The vertical cell values of the top-left cell of each cluster. This can also be a vector of values in case there are several clusters.

col [integerish(1)]:
The horizontal cell value of the top-left cell of each cluster. This can also be a vector of values in case there are several clusters.

width [integerish(1)]:
The width of each cluster in cells. This can also be a vector of values in case there are several clusters

height [integerish(1)]:
The height of each cluster in cells. This can also be a vector of values in case there are several clusters

id [character(1)]:
When data are clustered, it is often the case that the data are segregated according to one of the variables of interest. In such cases, this variable needs to be registered as cluster ID.

The slot header describes in which rows the header informations are stored and how they should be treated. It contains the properties

row [integerish(.)]
The rows in which the header information are stored.

rel [logical(1)]
Whether or not the values in row are relative to the cluster positions or whether they refer to the overall table.

merge [logical(1)]
When there is more than one row, this determines whether or not those rows should be merged.

The slot meta describes information concerning the values in a spreadsheet. It contains the properties

del [character(.)]
The symbol(s) that are used as delimiter in the table to reorganise.

dec [character(.)]
The symbol(s) that are used as decimal symbol in the table to reorganise.

na [character(.)]
The symbol(s) that are used as "not available" values in the table to reorganise.

Each element in the slot variables is a list that describes one of the variables that shall be comprised in the final database. Variables are either so-called identifying variables, which indicate observation units, or measured variables, which carry the observed values. Identifying variables are described by the properties:

type [character(1)]:
The value "id" signals that this is an identifying variable.

col [integerish(1)]:
The column(s) in which the variable values are recorded.

row [integerish(1)]:
The row(s) in which the variable values are recorded.

split [character(1)]:
A regular expression that matches the part of values that are supposed to be part of the variable, when the cells contain more than one value (for example separated by a ",").

dist [character(1)]:
Whether or not the variable is distinct from a cluster. This is the case when the variable is not systematically available for all clusters and thus needs to be registered separately from the clusters.

rel [logical(1)]:
Whether or not the values in row and col are relative to the cluster positions or whether they refer to the overall table.

Measured variables are described by the properties:

type [character(1)]:
the value "values" signals that this is a values variable.

unit [character(1)]:
The unit in which the values shall be recorded in the database.

factor [character(1)]:
A factor to transform the values to unit. For instance, if values are recorded in acres, but shall be contained in the database in hectare, the factor would be 0.40468.

row [integerish(1)]:
The row(s) in which the variable values are recorded.

col [integerish(1)]:
The column(s) in which the variable values are recorded.

rel [logical(1)]:
Whether or not the values in row and col are relative to the cluster positions or whether they refer to the overall table.

key [character(1)]:
If the variable is recorded, together with other variables, so that the variable names are listed in one column and the respective values are listed in another column, give here the name of the column that contains the variable names.

dist [character(1)]:
Whether or not the variable is distinct from a cluster. This is the case when the variable is not systematically available for all clusters and thus needs to be registered separately from the clusters.

value [character(1)]:
If the variable is recorded, together with other variables, so that the variable names are listed in one column and the respective values are listed in another column, give here the level in the names column that stands for the values of this variable.

The default schema description contains an example of an identifying and a values variables. Further identifying and values variables would be added when more variables are contained in a table.