Set Groups

This function allows to set groups for rows, columns or clusters that shall be summarised.

setGroups(schema = NULL, rows = NULL, columns = NULL)

Arguments

schema

[schema(1)]
In case this information is added to an already existing schema, provide that schema here (overwrites previous information).

rows

[list(3)]
the output of .sum indicating the rows and a function according to which those rows should be summarised.

columns

[list(3)]
the output of .sum indicating the columns and a function according to which those columns should be summarised.

Value

An object of class schema.

See also

Examples

# please check the vignette for examples