Reorganise a table

This function takes a disorganised messy table and rearranges columns and rows into a tidy table.

reorganise(input = NULL, schema = NULL)

Arguments

input
[data.frame(1)]
table to reorganise.

schema
[symbol(1)]
the schema description for reorganising input.

Value

A (tidy) table which is the result of employing schema on input.

Examples

# read in a disorganised messy dataset (without column names)
library(readr)
library(magrittr)
ds <- system.file("test_datasets", package = "tabshiftr")
input <- read_csv(file = paste0(ds, "/table_mismatch_3.csv"),
                  col_names = FALSE, col_types = cols(.default = "c"))
input

#> # A tibble: 13 x 7
#>    X1          X2        X3         X4          X5        X6         X7    
#>    <chr>       <chr>     <chr>      <chr>       <chr>     <chr>      <chr> 
#>  1 commodities harvested production <NA>        <NA>      <NA>       <NA>  
#>  2 unit 1      <NA>      <NA>       <NA>        <NA>      <NA>       <NA>  
#>  3 soybean     1111      1112       year 1      <NA>      <NA>       <NA>  
#>  4 maize       1121      1122       year 1      <NA>      <NA>       <NA>  
#>  5 soybean     1211      1212       year 2      <NA>      <NA>       <NA>  
#>  6 maize       1221      1222       year 2      <NA>      <NA>       <NA>  
#>  7 <NA>        <NA>      <NA>       <NA>        <NA>      <NA>       <NA>  
#>  8 commodities harvested production commodities harvested production <NA>  
#>  9 unit 2      <NA>      <NA>       unit 3      <NA>      <NA>       <NA>  
#> 10 soybean     2111      2112       soybean     3111      3112       year 1
#> 11 maize       2121      2122       maize       3121      3122       year 1
#> 12 soybean     2211      2212       soybean     3211      3212       year 2
#> 13 maize       2221      2222       maize       3221      3222       year 2

# put together schema description
schema <- setCluster(id = "territories", top = c(1, 8, 8), left = c(1, 1, 4),
                     width = 3, height = 6) %>%
  setHeader(rows = 1, relative = TRUE) %>%
  setIDVar(name = "territories", columns = 1, row = 2, relative = TRUE) %>%
  setIDVar(name = "year", columns = 4, row = c(3:6), distinct = TRUE) %>%
  setIDVar(name = "commodities", columns = 1, relative = TRUE) %>%
  setObsVar(name = "harvested", unit = "ha", columns = 2, relative = TRUE) %>%
  setObsVar(name = "production", unit = "t", columns = 3, relative = TRUE)

# get the tidy output
reorganise(input, schema)

#> # A tibble: 12 x 5
#>    territories year   commodities harvested production
#>    <chr>       <chr>  <chr>           <dbl>      <dbl>
#>  1 unit 1      year 1 maize            1121       1122
#>  2 unit 1      year 1 soybean          1111       1112
#>  3 unit 1      year 2 maize            1221       1222
#>  4 unit 1      year 2 soybean          1211       1212
#>  5 unit 2      year 1 maize            2121       2122
#>  6 unit 2      year 1 soybean          2111       2112
#>  7 unit 2      year 2 maize            2221       2222
#>  8 unit 2      year 2 soybean          2211       2212
#>  9 unit 3      year 1 maize            3121       3122
#> 10 unit 3      year 1 soybean          3111       3112
#> 11 unit 3      year 2 maize            3221       3222
#> 12 unit 3      year 2 soybean          3211       3212