Set an identifying variable

Identifying variables are those variables that describe the (qualitative) properties that make each observation (as described by the observed variables) unique.

setIDVar(
  schema = NULL,
  name = NULL,
  value = NULL,
  columns = NULL,
  row = NULL,
  split = NULL,
  merge = NULL,
  relative = FALSE,
  distinct = FALSE
)

Arguments

schema
[schema(1)]
In case this information is added to an already existing schema, provide that schema here (overwrites previous information).

name
[character(1)]
Name of the new identifying variable.

value
[character(1)]
In case the variable is an implicit variable (i.e., which is not in the origin table), specify it here.

columns
[integerish(.)]
The column(s) in which the values of the new variable are recorded.

row
[integerish(.)]
In case the variable is in several columns, specify here additionally the row in which the names are recorded.

split
[character(1)]
In case the variable is part of a compound value, this should be a regular expression that splits the respective value off of that compound value.

merge
[character(1)]
In case a variable is made up of several columns, this should be the character string that would connect the two columns (e.g., an empty space " ").

relative
[logical(1)]
whether or not the values provided in columns and row are relative to the cluster position(s) or whether they refer to the overall table.

distinct
[logical(1)]
Whether or not the variable is distinct from a cluster. This is the case when the variable is not systematically available for all clusters and thus needs to be registered separately from the clusters.

Value

An object of class schema.

Details

Please also take a look at the currently suggested strategy to set up a schema description.

See also